refactor: LottoNumber 제한조건 enum으로 추출
