+++ src/main/java/step2/LottoNumberAutoGenerationStrategy.java
@@ -1,5 +1,7 @@
 package step2;
 
+import step2.vo.LottoNumberValue;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -7,7 +9,12 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class LottoNumberAutoGenerationStrategy implements LottoNumberGenerationStrategy{
-    private final List<Integer> LOTTO_NUMBER_VALUES = IntStream.range(1, 46).boxed().sorted().collect(Collectors.toList());
+    private final List<Integer> LOTTO_NUMBER_VALUES =
+            IntStream.rangeClosed(LottoNumberValue.MIN.value, LottoNumberValue.MAX.value)
+            .boxed()
+            .sorted()
+            .collect(Collectors
+                    .toList());
 
     @Override
     public int generateNumber() {
