+++ src/main/java/step2/vo/LottoNumberValue.java
@@ -1,0 +1,18 @@
+package step2.vo;
+
+public enum LottoNumberValue {
+    MIN(1), MAX(45);
+
+    public int value;
+
+    LottoNumberValue(int value) {
+        this.value = value;
+    }
+
+    public static boolean isRange(int lottoNumber) {
+        if (lottoNumber >= MIN.value && lottoNumber <= MAX.value) {
+            return true;
+        }
+        return false;
+    }
+}
