+++ src/test/java/step2/WinTest.java
@@ -1,6 +1,7 @@
 package step2;
 
 import org.junit.jupiter.api.Test;
+import step2.vo.MatchNumber;
 
 import java.util.Arrays;
 import java.util.HashMap;
